% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Counter.R
\name{outlier_counter}
\alias{outlier_counter}
\title{Outlier Counter}
\usage{
outlier_counter(df)
}
\arguments{
\item{df}{The only parameter that 'Outlier Counter' takes is an arbitrary dataframe. Therefore intuitively I
named it as "df".}
}
\value{
The function will return a vector of characters which indicates either the number of outliers in
each column or "Not Numeric" if the corresponding column is not numeric. If the given parameter is
not a dataframe, it will return an error.
}
\description{
'Outlier Counter' will return the number of outliers in each numeric column of a given dataframe.
Here outlier is defined as any number above Q3+1.5IQR or below Q1-1.5IQR (NA will be omitted). If
a column is not numeric, then it will returns "Not Numeric". If the given parameter is not
dataframe, it will return an error.
}
\examples{
outlier_counter("ABC")
Error in outlier_counter("ABC") : Please pass a dataframe
outlier_counter(iris)
}
